backend/requirements.txt
django>=5.1,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
drf-spectacular>=0.28,<1
django-filter>=24.0,<25
reportlab>=4.1,<5
